+++ src/protocol/srt/srt_socket_service.cpp
@@ -453,7 +453,7 @@ namespace srt {
         srt_packet pkt;
         pkt.set_control_type(ack);
         pkt.set_type_information(ack_number);
-        pkt.set_timestamp(std::chrono::duration_cast<std::chrono::microseconds>(now - connect_point).count());
+        pkt.set_timestamp((uint32_t)std::chrono::duration_cast<std::chrono::microseconds>(now - connect_point).count());
         pkt.set_socket_id(get_sock_id());
         /// 添加到ack队列中
         _ack_queue_->add_ack(ack_number);
@@ -817,7 +817,7 @@ namespace srt {
         /// 更新上一次收到的时间
         last_receive_point = std::chrono::steady_clock::now();
         if (_packet_receive_rate_)
-            _packet_receive_rate_->update_receive_rate(buff->size() + 16);
+            _packet_receive_rate_->update_receive_rate((uint16_t)buff->size() + 16);
 
         if (srt_pkt->get_control()) {
             return handle_control(srt_pkt, buff);
@@ -854,7 +854,7 @@ namespace srt {
         /// 统计数据
         /// 丢入接收队列
         if (_packet_receive_rate_)
-            _packet_receive_rate_->update_estimated_capacity(pkt->get_packet_sequence_number(), buff->size() + 16, pkt->get_in_order() || pkt->is_retransmitted());
+            _packet_receive_rate_->update_estimated_capacity(pkt->get_packet_sequence_number(), (uint16_t)buff->size() + 16, pkt->get_in_order() || pkt->is_retransmitted());
         _receive_queue->input_packet(buff, pkt->get_packet_sequence_number(), pkt->get_time_stamp());
         if (srt_socket_base::report_nak && !report_nak_begin) {
             do_nak();
